Tests expecting failed results will throw invalid op_code
When writing unit tests for examples that should fail, refactor tests
such that the method that errors is the last method before being caught
by the error handler as an invalid operation code.

Depending on the Ethereum client used to run the tests (Parity dev,
testrpc, etc); errors may be handled differently. We're running our
tests against testrpc which catched invalid transactions by throwing an
error and catching in the promise chain.